Get rid of unnecessary calls to toDouble().
R=scheglov@google.com

Review URL: https://codereview.chromium.org//431713002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@38749 260f80e4-7a28-3924-810f-c04153c831b5

+++ pkg/analysis_server/test/integration/integration_tests.dart
@@ -709,8 +709,7 @@ class Server {
    * [debugStdio] has been called.
    */
   void _recordStdio(String line) {
-    double elapsedTime = _time.elapsedTicks.toDouble() /
-        _time.frequency.toDouble();
+    double elapsedTime = _time.elapsedTicks / _time.frequency;
     line = "$elapsedTime: $line";
     if (_debuggingStdio) {
       print(line);
